restructure reset students password
