+++ src/pages/teacherDashboard/classes/Classes.tsx
@@ -7,7 +7,7 @@ import ClassTable from "./ClassTable"
 import CreateClassForm from "./CreateClassForm"
 import JoinClassRequest from "./joinClassRequest/JoinClassRequest"
 import JoinClassRequestTable from "./JoinClassRequestTable"
-import ResetStudentsPassword from "./class/ResetStudentsPassword"
+import ResetStudentsPassword from "./resetStudentsPassword/ResetStudentsPassword"
 import { type RetrieveUserResult } from "../../../api/user"
 import TransferStudents from "./transferStudents/TransferStudents"
 import UpdateStudentUser from "./class/UpdateStudentUser"
